requirements/development.txt
-r base.txt

flake8==2.4.1
pylint==1.4.3
pylint-django==0.6
selenium==2.45.0
